fix: Move all js files into src
