+++ package.json
@@ -4,7 +4,7 @@
   "repository": "https://github.com/looky/looky-desktop",
   "license": "ISC",
   "version": "0.1.24",
-  "main": "electron/main.js",
+  "main": "dist/js/electron/main.js",
   "author": "Looky Labs, Inc.",
   "bin": {
     "release": "./scripts/release/index.js"
