+++ src/js/electron/extensions.js
